wip: try to fix things

+++ projects/rocprim/rocprim/include/rocprim/device/device_reduce_by_key.hpp
@@ -198,9 +198,9 @@ hipError_t reduce_by_key_impl_wrapped_config(void*                     temporary
             const unsigned int items_per_block = block_size * params.kernel_config.items_per_thread;
 
             const size_t size_limit = params.kernel_config.size_limit;
-            const size_t aligned_size_limit
-                = ::rocprim::max<size_t>(size_limit - (size_limit % items_per_block),
-                                         items_per_block);
+            const size_t aligned_size_limit = ::rocprim::max<size_t>(
+                size_limit - (size_limit % static_cast<std::size_t>(items_per_block)),
+                items_per_block);
 
             const size_t limited_size     = std::min<size_t>(size, aligned_size_limit);
             const bool   use_limited_size = limited_size == aligned_size_limit;
